+++ subprojects/maven/src/main/java/org/gradle/api/publication/maven/internal/action/MavenDeployAction.java
@@ -15,12 +15,9 @@
  */
 package org.gradle.api.publication.maven.internal.action;
 
-import org.apache.http.conn.HttpHostConnectException;
 import org.apache.maven.artifact.ant.RemoteRepository;
 import org.gradle.api.GradleException;
 import org.gradle.api.publish.maven.internal.publisher.MavenProjectIdentity;
-import org.gradle.internal.UncheckedException;
-import org.gradle.internal.resource.transport.http.HttpErrorStatusCodeException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.sonatype.aether.RepositorySystem;
@@ -33,10 +30,8 @@ import org.sonatype.aether.repository.Proxy;
 import org.sonatype.aether.util.repository.DefaultProxySelector;
 
 import java.io.File;
-import java.net.SocketTimeoutException;
 import java.util.Collection;
 import java.util.List;
-import java.util.concurrent.Callable;
 
 public class MavenDeployAction extends AbstractMavenPublishAction {
     private static final Logger LOGGER = LoggerFactory.getLogger(MavenDeployAction.class);
@@ -79,38 +74,15 @@ public class MavenDeployAction extends AbstractMavenPublishAction {
 
         LOGGER.info("Deploying to {}", gradleRepo.getUrl());
 
-        deployWithRetry(() -> { repositorySystem.deploy(session, request); return null; }, maxDeployAttempts, initialBackOff);
-    }
+        MavenDeployRetrier deployRetrier = new MavenDeployRetrier(
+            () -> { repositorySystem.deploy(session, request); return null; },
+            maxDeployAttempts,
+            initialBackOff,
+            remoteRepository
+        );
 
-    private void deployWithRetry(Callable operation, int maxDeployAttempts, int backoff) throws DeploymentException {
-        int retries = 0;
-        while (retries < maxDeployAttempts) {
-            retries++;
-            Throwable failure;
-            try {
-                operation.call();
-                if (retries > 1) {
-                    LOGGER.info("Successfully deployed resource after {} retries", retries - 1);
-                }
-                break;
-            } catch (Exception throwable) {
-                failure = throwable;
-            }
-            boolean doNotRetry = !isLikelyTransientNetworkingIssue(failure);
-            if (doNotRetry || retries == maxDeployAttempts) {
-                throw new DeploymentException("Could not deploy to remote repository | " + failure.getMessage(), failure);
-            } else {
-                LOGGER.info("Error while accessing remote repository {}. Waiting {}ms before next retry. {} retries left", remoteRepository, backoff, maxDeployAttempts - retries, failure);
-                try {
-                    Thread.sleep(backoff);
-                    backoff *= 2;
-                } catch (InterruptedException e) {
-                    throw UncheckedException.throwAsUncheckedException(e);
-                }
-            }
-        }
+        deployRetrier.deployWithRetry();
     }
-
 
     private org.sonatype.aether.repository.RemoteRepository createRepository(RemoteRepository gradleRepo) {
         org.sonatype.aether.repository.RemoteRepository repo = new org.sonatype.aether.repository.RemoteRepository("remote", gradleRepo.getLayout(), gradleRepo.getUrl());
@@ -131,18 +103,4 @@ public class MavenDeployAction extends AbstractMavenPublishAction {
         return repo;
     }
 
-    private static boolean isLikelyTransientNetworkingIssue(Throwable failure) {
-        if (failure instanceof SocketTimeoutException || failure instanceof HttpHostConnectException) {
-            return true;
-        }
-        if (failure instanceof HttpErrorStatusCodeException) {
-            HttpErrorStatusCodeException httpError = (HttpErrorStatusCodeException) failure;
-            return httpError.isServerError();
-        }
-        Throwable cause = failure.getCause();
-        if (cause != null && cause != failure) {
-            return isLikelyTransientNetworkingIssue(cause);
-        }
-        return false;
-    }
 }
